reorganize tests with injection
into tests/with-injection
