--- tests/lib-cli-command/object/lib-cli-command-object-text.test.js
@@ -1,5 +1,0 @@
-const command = require('../../../lib/cli-command.js');
-
-test('cli-command object', () => {
-  expect(command).toMatchSnapshot();
-});
