--- tests/lib-example-defaults/lib-example-defaults.test.js
@@ -1,17 +1,0 @@
-const lib = require('../../lib/lib.js');
-
-const ioMocks = require('../helpers/io-mocks.js');
-
-test('create alice-bobbi module with defaults', () => {
-  const mysnap = [];
-
-  const mocks = ioMocks(mysnap);
-
-  const options = {
-    name: 'alice-bobbi',
-    generateExample: true,
-  };
-
-  return lib(options, mocks)
-    .then(() => { expect(mysnap).toMatchSnapshot(); });
-});
