--- tests/lib-view-with-example-defaults/lib-view-with-example-defaults.test.js
@@ -1,18 +1,0 @@
-const lib = require('../../lib/lib.js');
-
-const ioMocks = require('../helpers/io-mocks.js');
-
-test('create alice-bobbi view module with example, with defaults', () => {
-  const mysnap = [];
-
-  const mocks = ioMocks(mysnap);
-
-  const options = {
-    name: 'alice-bobbi',
-    generateExample: true,
-    view: true,
-  };
-
-  return lib(options, mocks)
-    .then(() => { expect(mysnap).toMatchSnapshot(); });
-});
