+++ tests/with-injection/cli/command/object/lib-cli-command-object-text.test.js
@@ -1,0 +1,5 @@
+const command = require('../../../../../lib/cli-command.js');
+
+test('cli-command object', () => {
+  expect(command).toMatchSnapshot();
+});
