+++ tests/with-injection/create/view/with-defaults/create-view-with-defaults.test.js
@@ -1,0 +1,17 @@
+const lib = require('../../../../../lib/lib.js');
+
+const ioMocks = require('../../../helpers/io-inject.js');
+
+test('create alice-bobbi view module with defaults', () => {
+  const mysnap = [];
+
+  const mocks = ioMocks(mysnap);
+
+  const options = {
+    name: 'alice-bobbi',
+    view: true,
+  };
+
+  return lib(options, mocks)
+    .then(() => { expect(mysnap).toMatchSnapshot(); });
+});
